authentication and authorization functionality

+++ src/main/java/com/project/stayEase/service/HotelServiceImpl.java
@@ -7,11 +7,14 @@ import com.project.stayEase.dto.HotelResponseDto;
 import com.project.stayEase.dto.RoomResponseDto;
 import com.project.stayEase.entity.Hotel;
 import com.project.stayEase.entity.Room;
+import com.project.stayEase.entity.User;
 import com.project.stayEase.repository.HotelRepository;
+import com.project.stayEase.security.SecurityUtils;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -27,6 +30,7 @@ public class HotelServiceImpl implements HotelService{
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
+    private final SecurityUtils securityUtils;
 
 
     @Override
@@ -34,6 +38,7 @@ public class HotelServiceImpl implements HotelService{
 
         log.info("creating new hotel with hotelRequestDto={}", hotelRequestDto.getName());
         Hotel hotel = modelMapper.map(hotelRequestDto, Hotel.class);
+        hotel.setOwner(securityUtils.getCurrentuser());
         hotelRepository.save(hotel);
         return modelMapper.map(hotel, HotelResponseDto.class);
 
@@ -43,6 +48,9 @@ public class HotelServiceImpl implements HotelService{
     public HotelResponseDto getHotelById(Long id) {
         log.info("getting hotel with id={}", id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with id " + id));
+        if(!hotel.getOwner().getId().equals(securityUtils.getCurrentUserId())){
+            throw new AccessDeniedException("You are not access this hotel");
+        }
         return modelMapper.map(hotel, HotelResponseDto.class);
     }
 
@@ -50,6 +58,9 @@ public class HotelServiceImpl implements HotelService{
     public HotelResponseDto updateHotelById(Long id, HotelRequestDto hotelRequestDto) {
         log.info("updating hotel with id={}", id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with id " + id));
+        if(!hotel.getOwner().getId().equals(securityUtils.getCurrentUserId())){
+            throw new AccessDeniedException("You are not allowed to modify this hotel");
+        }
         modelMapper.typeMap(HotelRequestDto.class, Hotel.class)
                 .addMappings(mapper -> {
                     mapper.skip(Hotel::setAmenities);
@@ -64,8 +75,10 @@ public class HotelServiceImpl implements HotelService{
     @Override
     public void activateHotel(Long id) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with id " + id));
+        if(!hotel.getOwner().getId().equals(securityUtils.getCurrentUserId())){
+            throw new AccessDeniedException("You are not access this hotel");
+        }
         hotel.setActive(true);
-//        hotelRepository.save(hotel);
         for(Room room: hotel.getRooms()){
             inventoryService.initializeRoomForHalfYear(room);
         }
@@ -76,7 +89,9 @@ public class HotelServiceImpl implements HotelService{
     @Transactional
     public void deleteHotelById(Long id) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with id " + id));
-
+        if(!hotel.getOwner().getId().equals(securityUtils.getCurrentUserId())){
+            throw new AccessDeniedException("You are not allowed to delete this hotel");
+        }
         for(Room room: hotel.getRooms()){
             inventoryService.deleteAllInventoriesForRoom(room);
         }
